requirements.txt
selenium~=4.34.0
pytest~=8.4.1
pytest-html==4.1.1
allure-pytest==2.14.0
python-dotenv==1.0.1
faker==24.11.0
allure-python-commons~=2.14.3